Add release notes for package manager
Just to keep track of where minikube is being served from besides the
places we explicitly own.  Added the brew formula as well as the arch
user repository.

AFAIK we can't do much about the Arch AUR besides flag it for update,
but we can submit pull requests for new brew formulas through the
homebrew-core repository.