Pass in k8s ldflags to minikube also
We are using k8s pkg/version in minikube cmd to print out the
default version.  This was falling back to the defaults in pkg/version/base.go which were an ugly format string.  This wasn't actually being passed into cluster.go because we didn't pass the string in if was a default version.