Add LocalClient to bypass RPC for drivers
This adds a new libmachine API implementation that skips RPC and instead
uses a vendored in driver directly.  This removes the need to download a
separate driver binary.  Currently, only the "core" drivers, and kvm are
supported as vendored in.  This means that xhyve is the only driver that
requires a separate binary for now.