test: check stopped status repeatedly
In TestStartStop(), check for stopped status once in 5 seconds,
up to 30 seconds, instead of always sleeping for 30 seconds
before stopping it. That way we can reduce duration of the test.

To do that, we need to split out MinikubeRunner.CheckStatus() into
CheckStatusNoFail() that doesn't lead to T.Fatalf(). Other call sites
of CheckStatus() would not be then affected.