Detect and test against ISO changes
When changes are made to the ISO folder, rebuild and upload the newly
modified image.  Use this uploaded image as the default in the e2e test
binaries, so that the e2e test suite runs against the PR's changes.

TODO: Subsequent PRs (no changes to ISO) will still run against the
default version in the makefile.  We should figure out if we want to
always run tests against the "latest" image.