Add minikube/releases/stable.txt for downloading
The latest stable kubectl can be used for downloading with the script

`curl -LO
https://storage.googleapis.com/kubernetes-release/release/$(curl -s
https://storage.googleapis.com/kubernetes-release/release/stable.txt)/bin/linux/amd64/kubectl`

We should allow minikube to be downloaded in a similar manner.  Right
now, downstream docs and demos need to hardcode the version.  This adds
gs://minikube/releases/stable.txt

With this PR, you will be able to reference

`curl -LO https://storage.googleapis.com/minikube/releases/$(curl -s
https://storage.googleapis.com/minikube/releases/stable.txt)/minikube-linux-amd64`

It also changes the notification update on new releases to just parse
this file instead of the releases.json, as it will be much more
straightforward.