Update linux libvirt-bin dependency
Since we compile with CGO on linux now, there is a dependency on
libvirt-bin, since minikube is now dynamically linked to libvirt.so.
This is unfortunate if you're using virtualbox on linux, but should
require no changes if you're using KVM already.

For better UX, we should probably just default linux to KVM for now,
until we have a no-vm driver.