minikube-iso: Use buildroot branch 2017-02
This rebases minikube-iso to the release 2017-02.

A change in buildroot required the introduction of a `external.desc` as
multiple external trees are supported in modern buildroot versions. This
required change is described here [0].

[0] https://buildroot.org/downloads/manual/manual.html#br2-external-converting

The existing isolinux.cfg was not included in the config and thus
ignored, include it.

The 4.9.13 kernel included in buildroot 2017-02 is not compatible with
VBox guest additions 5.1.6, thus bump guest additions to 5.1.18.

The defconfig of 4.9.13 has been cleaned up.

systemd 232 included in buildroot 2017-02 is incompatible with docker <
17.03 without boot parameter systemd.legacy_systemd_cgroup_controller=yes

Signed-off-by: Thomas Graf <tgraf@suug.ch>