Fix godep-* hack scripts for 1.7
* Added a godep-update-k8s script for upgrading.
$ KUBE_VERSION ./hack/godeps/godep-update-k8s.sh

* Added client-go version ldflags to localkube and minikube.  The
version package is now split in both kubernetes and client-go, so we
need to supply both ldflags.

* Change cross builds to use BUILD_IN_DOCKER

* Add a patch for kube-proxy
The setup of kube-proxy recently changed, making the registering of the
component config a fatal error instead of a warning.  This patch
reverts it to be a warning, which is a no-op for us anyways since we
don't supply a config file.