Remove 30 second wait from start stop
I removed this awhile ago, but I'm not sure how it snuck back in.
The original issue was that the VM was getting a new IP, but the driver
was parsing the leases file and still picking the old IP.  So the only
way to fix it was to wait for the old lease to expire.