Add target for none driver integration
Theres currently no good way to run the integration tests for the none
driver locally.  Now that we've added the e2e target, we can depend on
that.  Running the e2e test binary locally requires the testdata folder
to be in a different path, since it is relative to where the code is
being ran.  I added a testdata-dir flag to the integration tests so we
can set it appropriately for when we want to run the e2e binary by
itself.

I also made the e2e test binary a PHONY target, so we rebuild it each
time.  I had trouble collecting the dependencies on the integration
test packages