Concurrent provisioning and localkube setup
After the machine has been created and started, we can perform the
docker setup at the same time as the localkube setup.  Specifically,
once the machine is ssh-able, we can start moving files while docker is
getting started.

Since we don't have anything directly exposed through the libmachine api
to know when we can do this, we create an event channel on which our
libmachine create step publishes events as steps complete.  Once we
receive an event that the machine has been created, started, ssh-able,
and have saved a configuration file for it, we can start the localkube
provisioning.