cni: Configure portmap plugin
Previously CNI was configured to use the bridge plugin.
But to support nodePorts, the portmap plugin had to be configued as well.

With this change the CNI config is moved to a config list (as supported
by cni >= 0.3.0). In this config list there is now the bridge configuration
and an additional one for the portmap plugin.

If this config is used on minikube, then the ingress pod will also work
with CNI.

Fixes #1987