linux 4.16, buildroot 2018.05
linux is upgraded to 4.16, binutils is upgraded to 2.30, and
libelf is installed for the newer kernel.

Other configuration changes are changes in defaults and ordering.

Now using glibc 2.27. systemd 237, and Go 1.10.

vbox-guest is removed because it's upstream as of linux 4.16.

More on buildroot at https://git.buildroot.net/buildroot/tree/CHANGES