pkg.drivers.none: refactor commands
Refactor Kill(), Restart(), Stop(), Remove() to use runCommand() consistently,
not duplicate `sudo`, handle errors consistently as well.

In Restart(), `exec.Command()` was used to run bash code,
which is plain incorrect.

In Remove(), local state is wiped before stopping containers, which is
ugly.

For Restart() and Kill(), I'm not sure which semantics are expected
because they are not used and seemingly are implemented only for
compatibility with "github.com/docker/machine/libmachine/drivers".Driver().
Still:
- for Kill(), removing minikube state but not killing containers doesn't
  sound right
- for Restart(), docker-machine docs do not say anything about
  try-restarting, and restarting kubelet does not stop containers