ensure that /dev has settled before operating
Minikube automount relies on a VM disk to be mounted (usually /dev/sda)
to provide .ssh credentials.

This commit adds a dependency on systemd-udev-settle to ensure that all devices
(including disks) are up and available before executing the mount
script.

We use systemd-udev-settle
https://github.com/systemd/systemd/blob/master/units/systemd-udev-settle.service.in

To ensure this. Note that in the above link, we can read a comment
stating that systemd-udev-settle is used to ensure a populated /dev
during bootup, which is exaclty what we want.

Signed-off-by: David Genest <david.genest@gmail.com>