Rework version handling to be less brittle and more consistent.
Version was generally being passed around as a string, I suspect due to
the use of JSON as a configuration store and wanting it to map to certain
configuration structures. This unfortunately led to issues where it was
implicitly required that all Kubernetes versions be specified with a
leading 'v' prefix or various operations would fail.

This patch:
- Replaces as many instances of string-based versions with semver.Version
  objects as possible.
- Makes version parsing more reliable by consistently using
  strings.TrimPrefix to remove the 'v' prefix when needed.
- Kubernetes versions stored in profiles are normalised to not having the
  'v' prefix. Existing profile configurations with a 'v' prefix will still
  work.
- Validates versions being passed in to make sure they'll parse.

Fixes #3253

Remove tests that are no longer needed.

Fix templates.