Code review comments
Instead of checking in default config.toml, save it at /tmp/config.toml
on the node upon enable and copy it back upon disable.

Also, instead of using the prestop hook, intercept the SIGTERM kill signal upon pod
termination, disable gvisor, and then exit with code 0. This should work
better because now we will be able to see the logs from disabling, and
because the prestop hook wouldn't consistenly run the disable command
and clean up the pod correctly.