Change restart policy on gvisor pod
Change the restart policy on the gvisor pod to Always. This way, if a
user runs

minikube addons enable gvisor
minikube stop
minikube start

when the addon manager tries to restart the gvisor pod, it will be
restarted and gvisor will start running automatically. This PR also adds an
integration test for this functionality.