Do not wait for k8s-app pods when starting with CNI
Previously, minikube was waiting for some pods labeled with
"k8s-app" (e.g. "kube-dns") to be ready before indicating that
"minikube start" was successful. However, in the case of
"network-plugin=cni", network of such pods is managed by a CNI
plugin which is usually started after "minikube start" has returned
making the waiting for the pods to fail with:

    Error starting cluster: wait: waiting for k8s-app=kube-dns: timed out
    waiting for the condition

This commit omits waiting for "k8s-app" pods if minikube is started
with "network-plugin=cni".

Reported-by: Michal Rostecki <mrostecki@opensuse.org>
Signed-off-by: Martynas Pumputis <m@lambda.lt>