Reenable systemd-timesyncd, except on virtualbox
This commit attempts to add back the missing time synchronization feature
to Minikube that was removed earlier with #3476.

As mentioned in #1378 we have an alternative solution for time
synchronization for Oracle VirtualBox, so there we don't want to enable
systemd-timesyncd.

We are using systemd conditional activation on systemd-timesyncd and
exclude systems that have an oracle hypervisor hosting the vm (currently
that's virtualbox for our purposes).