Add note for bare-metal driver
I open an issue (#5308) while trying the command to create the second local cluster.  It turns out that this operation isn't supported with a bare-metal (none) driver and the code is being changed to catch this restriction early.

I thought it might be nice to add a warning to the examples page as well.