Upgrade Buildroot to 2019.02 and VirtualBox to 5.2
New major version of Buildroot, with new Linux kernel.
Both are LTS versions, and are supported until 2020.

Buildroot is upgraded from 2018.05 to 2019.02 LTS,
and Linux from 4.15 to 4.19 LTS as included in BR2.

Also upgrade VirtualBox to the supported 5.2 version
of the older branch, so that it will work everywhere.

Need to remove the legacy symlink for crio.sock from
the image, since it was causing build time issues...