Use faster build for the default minikube target
When we are doing the "cross" compile targets, we
force rebuilding all packages (up-to-date or not).

For the default native target "out/minikube", we
can do a much faster build (10x!), for development.

Leaving all the cross target to rebuild, as before.

Also the _test.go files are only used for unit tests.