Store kubernetes clientset when waiting for apiserver
While looking at pprof data, I noticed that the line:

```go
client, err = k.client(k8s)
```

takes around .9 seconds to complete. Previously, we were calling it on
every iteration of the waitFor loop. Instead, if we store the clienset once
we have it, we should be able to speed up this function.

During local testing, this optimization reduced time spent waiting for
the apiserver by 13 seconds on average.