Don't use bash as the entrypoint for docker
The bash shell is not very good at playing init(1), let
tini handle this instead by doing `docker run --init`.

This makes it forward signals properly to e.g. make,
and also reaps processes for long-running containers.

However, we _need_ to override the entrypoint of the
karalabe/xgo image, since it'll self-destruct otherwise.

It (/build.sh) expects the command to be the import path
and cannot handle getting the bash command line to run.