Switched to using kubectl apply and kubectl delete for addons
kubectl apply with --prune wasn't working because it requires the
--namespace flag to be passed in and we have multiple namespaces for
addons.

We tried to split up addons within /etc/kubernetes/addons by
namespace, but this didn't work when disabling dashboard addon because
there were no files left in the kubernetes-dashboard subdirectory. So,
kubectl apply complained when trying to prune dashboard because no files
were being passed in.

So, I ended up removing all dependencies on prune and instead running
`kubectl apply` when enabling an addon and running `kubectl delete` when
disabling an addon.