Fail fast if waiting for SSH to be available
Integration tests were running long because we'd wait up to 22 minutes for SSH to succeed while creating the host. This commit introduces a two minute timer on creating the host. On my Macbook Pro with the docker driver, creating the host takes 5 seconds, so 2 minutes should be more than enough time.

Refer to #6608