Don't set localhost as ip on linux
On linux, we want to keep the IP set to the internal IP. This is because this IP is exposed and so we can use it on Linux.

This doesn't fix the IP issue on mac or windows, it's just a temporary fix for the linux case. We will need to find a way to expose internal ports on Mac/Windows so that the registry addon works as expected on those platforms.