Fix StartStopTest/group/containerd flake
This test was failing regularly with the following error showing up on all pods:

```
FailedCreatePodSandbox
```

after doing some digging, it seems that this error is always CNI related. I found these docs on the k8s website about running containerd in minikube:

https://kubernetes.io/docs/setup/learning-environment/minikube/

which says the flags `--network-plugin=cni` and `--enable-default-cni` should be set with running containerd. I tried passing those flags in, and ran this test locally 3x. It passed everytime, so I'm hoping this fixes the flake.