Move status command to TestFunctional/serial
I noticed that TestComponentHealth/parallel/ComponentHealth was failing with this error:

```
Error apiserver status: https://172.17.0.3:8441/healthz returned error 500:
[-]etcd failed: reason withheld
```

but by the time post mortem logs were printed the etcd container was up and running.

I think this test occasionally fails because apiserver healthz is not yet returning a 200 status when we run the test. We wait for healthz to return 200 on regular start, but not on soft start, which we run in `TestFunctional`.

This PR adds a retry, which should give the apiserver time to become healthy.