Verify that the build directory for buildroot is case sensitive
buildroot fails when the filesystem is just case-preserving, which
is the default for macOS filesystems. So it is better to fail
immediately than after spending 90min building things.

This commit also includes a `make case-sensitive-buildroot-dir` target
(implemented on macOS only) that creates a case-sensitive spareimage
and mounts it at the buildroot dir location. Any files from the previous
buildroot dir are moved to this new sparseimage.

I do not understand why the `hdiutil detach` in the `clean` target
needs the -force option when running via `make`. Otherwise it will
claim "resource busy". This does not happen when running directly
from the shell.