Update kubectl syntax to minikube syntax
Users of this documentation are likely setting up minikube without the kubectl command available on their local environment. Other issues have been opened regarding this in the past. I haven't added further instructions myself to line 146 but it may be helpful to express the difference between a standalone kubectl command and the double-dashes used by minikube to stop parsing flags. 

The changes made continue to make use of minikube through the remainder of the page instead of using kubectl directly.