add polkit check for kvm2 driver
When using Fedora Silverblue 34 with libvirt installed in the ostree
overlay, the group memberships aren't setup as they would be in a
traditional GNU/Linux system. However, if a polkit policy is in place
that allows the user access to the "org.libvirt.unix.manage" action then
the user is able to create virtual machines via libvirt regardless of
group membership.

Signed-off-by: Adam Miller <maxamillion@gmail.com>
Signed-off-by: Adam Miller <maxamillion@fedoraproject.org>