Add support for iscsi libraries in minikube build.
This is required for Longhorn as it expects the iscsi libraries and kernel modules
to be available. This patch adds the open-iscsi (and its required open-isns) to
provide the iscsiadm, iscsid and related binaries into thes ystem.

This patch also updates the minikube-automount script to mount /var/lib/longhorn
which is where all the replicas for longhorn are mounted. This is done to retain
those directories after a restart of minikube