Fix typo in help text for minikube docker-env
In the help text for the `minikube docker-env` command the suggestions for setting the docker command for Windows (cmd and powershell) don't work as output because if you don't pass the `--shell` flag, it adds a line at the beginning that will cause the cmd example to fail, and doesn't specify that you need to set the flag `--shell powershell` for it to work properly. I've updated the help text to provide a command that will work as expected for both shells.