remove hyperkit version checks
Minikube was using an undocumented API. The hyperkit developers
never promised a stable format for their version string.  In fact,
the version has changed from date-based to commit-based version
strings.

Since it has been several releases since the last issue caused by
older versions of hyperkit, and removal of obsolete code is often
something to be celebrated, this removes any checks for a minimum
compatible version of hyperkit.  If hyperkit is installed, it is
assumed to be a compatible version.

Signed-off-by: Pete Nelson <petiepooo@gmail.com>