Fix wrong use of loop variable in parallel test
This fixes an occurrence of a loop variable being incorrectly captured
in a parallel test (`TestVersionIsBetween`). With the previous code,
only the last test case is actually exercised. To work around this
problem, we create a local coapy of the range variable before the
parallel test, as suggested in the documentation for the `testing`
package:

https://pkg.go.dev/testing#hdr-Subtests_and_Sub_benchmarks

Some assertions had to be changed as they were incorrect.

Issue was found using the `loopvarcapture` linter.