Simplify containerd runtime configuration for gVisor
The previous code had a copy of `config.toml` which is easy to get out of
sync. Instead, append the necessary configuration to the existing file
to retain the rest of the configuration.

Also changed the location where `runsc` is downloaded to use the latest
release instead of hardcondig to a release which gets old quickly.