Adds check for volume existence in oci driver's createVolume()
As the function's description states:
It should not return err or change labels if volume already exists..

Info that we found a volume might be helpful tho..
As unlikely as it may sound.. user might create a minikube volume
and wonder why its actual minikube cluster is not starting.

We're deleteing TODO msg.
The issue was already closed.