Adds distinction between content/distribution digests
Being pkg/drivers/kic/types.go the source of truth for the version of
the container we're using to instantiate our kübernetes cluster in,
the pr should start here..

Initially I thought about hardcoding the contentDigest(a.k.a. imageId)
here as well, to then use it to check against the images inside the
kicDriver.. It later took another turn(we're retrieving it from tar).

Plus a collaborator showed me that it was a bad idea.. maintaining it
here would bean bumping it as part of the image build process.

The idea is based on the following concepts:

.contentDigest is the most reliable way to address image content:
if the image is tampered with after push to a registry,
the contentDigest we'd see after pull,
would be different than the one hardcoded here.
It is also part of the image itself, i.e. part of the tar archive;
thus giving us a way to always know if the cache is up to date,
even offline.

.distributionDigest is the most reliable way to determine which
image we're looking to pull from a registry; a tag can be detached
from an image and recycled, referencing another one, with different
content.
It is not part of the image itself; it is computed on the image in
compressed state.. and since different engines/mechanisms could use
different types of compression, this digest is totally unreliable
as a way to address content.

[*] refs:
https://windsock.io/explaining-docker-image-ids/
google/go-containerregistry#895 (comment)
https://stackoverflow.com/questions/45533005/why-digests-are-different-depend-on-registry
https://blog.aquasec.com/docker-image-tags -- follow links