Moves runtime/runner initialization inside PreCreateCheck
This is definitely messy.
but since the Driver methods have all pointer receivers..
We could move the initialization logic for the runner/runtime,
inside the first step of the (*LocalClient).Create call,
which is h.d.PrecreateCheck();
and be assured that the future calls to the none (*Driver),
will have them initialized.

To allow initialization outside of the registry,
we mimic what the Kic driver does, and move the config
of the driver, inside the driver itself.

Why is this messy?
This initialization thing, for clarity, should be explicitly carried
in a 'configure' step.. the future maintaner of this would cuss at me
for sure.
But since the (*LocalClient), docker-machine api implementation is
not going to be refactored soon...