Removes external lib dependency
Since the test is failing on architectures/platform I can't reproduce
on.. I'm readapting the test in such a way to minimize the amount of
code that can be responsible for the error.

We're not looking for all system processes in order to find our
mount processes(ps -ef); we're looking directly at the mount
processes that have been spawned by our test instead.

To say that a process is no longer alive:
We try to wait for conclusion to get the processState.
If the wait results in an *ExitError, it means that either
the proc has failed to run, or it exited failing(see doc);
in both cases we're happy.. since we previously made a check
to know that those same processes were alive.