no date in version string implies newer hyperkit
Really really REALLY old versions of hyperkit did not have a date in the version string.  The newest versions of hyperkit again do not have a date in the version string.
This changes the date check such that, instead of assuming hyperkit versions without a date are extremely old, it assumes they are new.  Without this, minikubes only works with hyperkit versions released on or after 20190201 but before 20210107 (aka version 40cbd5), failing for the most recent versions.  With this, it should now work with any version released on or after 20190201, which is when the fix that allowed it to work was committed.