moves runner/cruntime into libmachine && Ties Runner to Driver
It moves the following pieces into the libmachine:
 + command_runner thingy
 + cruntime thingy
 + sshutils renamed into pkg/libmachine/utils
 + !! Adds the responsibility to get the runner
      to the driver itself

TODO:
I'm not yet sure if the driver should export runner.Runner
methods or not... but I'm sure that the Driver should be
responsible to the initialization of the command *Runner*.

By having the Driver responsible to get the Runner,
we can toss the GetSSSHwhatever() logic from the driver
interface, that is mainly used to establish the ssh connection
as a mean to run commands inside the linux machine;
this is a design issue for the none(BareMetal) driver
implementation, and is definitely limitative for other drivers as
well.
If the Driver is responsible to get the Runner,
then we can use, through the driver, whatever mechanism
to run commands inside the linux machine.