Use the stable path for brew installed qemu firmware in machine config
Previously, machine config had hard coded the homebrew cellar path with
the patch version of qemu available at the time of machine creation,
and was prone to breaking anytime qemu was updated and the machine is
restarted. This change replaces the firmware path to use the prefix as
reported by 'brew --prefix qemu' (e.g. /opt/homebrew/opt/qemu on M1)
which is a symlink the the latest installed version of qemu
(e.g. /opt/homebrew/Cellar/qemu/<x.y.z>)