Base ISO cache file name on the full ISO URL
Currently, only the final component of the URL path is used. This means
that minikube may use the wrong cached image after the user changes the
ISO URL to another with the same final component.

Fix this by suffixing the cache file name with a hash of the entire URL.

Signed-off-by: Alberto Faria <afaria@redhat.com>